Fix tidy warnings & implement script

+++ src/DistributedPanelAnimation.cpp
@@ -9,7 +9,7 @@ DistributedPanelAnimation::DistributedPanelAnimation(Config* config) {
     // This Cycle is for how fast each panel iterates over the gradient
     // TODO: duration_ticks to Config
     this->cycle = new Cycle(30, false, UP_ONLY_CYCLE);
-    this->panel_distribution_ratio = (float) 1 / (float) config->number_of_panels;
+    this->panel_distribution_ratio = 1.0f / static_cast<float>(config->number_of_panels);
 }
 
 void DistributedPanelAnimation::start() {
@@ -27,7 +27,7 @@ void DistributedPanelAnimation::stop() {
 }
 
 float DistributedPanelAnimation::get_value_for_panel(int panel_index) const {
-    float value = this->cycle->current_value + ((float) panel_index * this->panel_distribution_ratio);
+    float value = this->cycle->current_value + (static_cast<float>(panel_index) * this->panel_distribution_ratio);
     if (value > 1) {
         return value - 1;
     }
